fix: rename api files, based on #59
