+++ src/pages/CartPage/CartListItem.tsx
@@ -5,7 +5,7 @@ import { formatNameForURL, formatPrice } from '@/helpers/helpersFunc';
 import { QuantityPiecesCart } from '@/components/QuantityPieces/QuantityPiecesCart';
 import { removeProductsFromCartAll } from '@/reducers/controller';
 import { useAnimations } from '@/components/CustomHook/AnimationsHook';
-import { useGetProductsQuery } from '@/api/ProductsAPI';
+import { useGetProductsQuery } from '@/api/_productsAPI';
 import { useMyURLContext } from '@/context/URLContext';
 
 export function CartListItem({
