+++ src/pages/ProductPage/ProductPage.tsx
@@ -8,7 +8,7 @@ import { addProductsToCart } from '@/reducers/controller';
 import { useSelector } from 'react-redux';
 import { useAnimations } from '@/components/CustomHook/AnimationsHook';
 import { ArrowBack } from '@/components/ArrowBack/ArrowBack';
-import { useGetProductQuery, useGetProductsQuery } from '@/api/ProductsAPI';
+import { useGetProductQuery, useGetProductsQuery } from '@/api/_productsAPI';
 import { Preloader } from '@/components/Preloader/Preloader';
 import { useMyURLContext } from '@/context/URLContext';
